refactor:extract getCanFiscalYear.js to api/
